Made the slide container into position relative
The plan is to use the relative positioning to
manually shift over to the next image by moving
the slide container over to the right, or left
by an amount. I made all the images in the
carousel the same size, so the position shift
should be similar. Also will by default, shift to
the right every 5 seconds.