Coded nextSlide and prevSlide logic in SlideShift
Coded the nextSlide and prevSlide logic in the
SlideShift.js module. It gets the current position
of the "right" style attribute and adds 800px
to it if we are moving to the nextSlide() or
-800px if we are going to the prevSlide().
Changed the expression to type cast the extracted
string from the attribute value into a Number
to prevent type conflicts with addition.