Finished nav buttons and timeout
Finished the image carousel and successfully
added the nav buttons with the stylings and the
click events to navigate to the specific image.
The setInterval will invoke the function given
after the set amount of miliseconds given, and
the nav dots will change their background-color
to red whenever the images change. I separated
it into 2 functions, one function for directly
clicking on the nav buttons, and one function
that will change the nav button's styling
to indicate the current image when the next and
previous buttons are used.